backend-flask/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
twilio==8.2.0
flask-limiter==3.5.0
python-jose==3.3.0
